Refactor name to proper name classs
