+++ fizzbuzz/src/test/java/edu/uagrm/main/FizzBuzzTest.java
@@ -4,51 +4,51 @@ import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-public class MainTest {
+public class FizzBuzzTest {
 
     @Test
     public void testInitShouldReturnsTrue() {
-        Main main = new Main();
-        assertTrue(main.init());
+        FizzBuzz fizzBuzz = new FizzBuzz();
+        assertTrue(fizzBuzz.init());
     }
 
     @Test
     public void testGetFizzBuzzIfNumberIsDivisibleByThreeReturnFizz(){
-        Main main = new Main();
+        FizzBuzz fizzBuzz = new FizzBuzz();
         String expectedValue = "Fizz";
-        String result = main.eval(3);
+        String result = fizzBuzz.eval(3);
         assertEquals(expectedValue, result);
     }
 
     @Test
     public void testGetFizzBuzzIfNumberIsDivisibleByFiveReturnBuzz(){
-        Main main = new Main();
+        FizzBuzz fizzBuzz = new FizzBuzz();
         String expectedValue = "Buzz";
-        String result = main.eval(5);
+        String result = fizzBuzz.eval(5);
         assertEquals(expectedValue, result);
     }
 
     @Test
     public void testGetFizzBuzzIfNumberIsDivisibleByFiveAndThreeReturnFizzBuzz(){
-        Main main = new Main();
+        FizzBuzz fizzBuzz = new FizzBuzz();
         String expectedValue = "FizzBuzz";
-        String result = main.eval(15);
+        String result = fizzBuzz.eval(15);
         assertEquals(expectedValue, result);
     }
 
     @Test
     public void testGetFizzBuzzIfNumberIsNoTDivisibleByFiveAndThreeReturnSameNumber(){
-        Main main = new Main();
+        FizzBuzz fizzBuzz = new FizzBuzz();
         String expectedValue = "1";
-        String result = main.eval(1);
+        String result = fizzBuzz.eval(1);
         assertEquals(expectedValue, result);
     }
 
     @Test
     public void testGetFizzBuzzIfNumberIsZeroReturnZero(){
-        Main main = new Main();
+        FizzBuzz fizzBuzz = new FizzBuzz();
         String expectedValue = "0";
-        String result = main.eval(0);
+        String result = fizzBuzz.eval(0);
         assertEquals(expectedValue, result);
     }
